+++ skin-api/src/main/java/com/personal/skin_api/product/service/dto/request/ProductRegisterServiceRequest.java
@@ -10,12 +10,13 @@ import org.springframework.web.multipart.MultipartFile;
 public class ProductRegisterServiceRequest {
     private final String productName;
     private final String productContent;
-    private final int price;
+    private final Long price;
     private final MultipartFile file;
     private final String email;
 
     @Builder
-    private ProductRegisterServiceRequest(final String productName, final String productContent, final int price, final MultipartFile file, final String email) {
+    private ProductRegisterServiceRequest(final String productName, final String productContent,
+                                          final Long price, final MultipartFile file, final String email) {
         this.productName = productName;
         this.productContent = productContent;
         this.price = price;
